:recycle: refactors the pages file to only containe internal paths
