+++ .gitignore
@@ -59,6 +59,7 @@ typings/
 
 # next.js build output
 .next
+out/
 
 # storybook build
 storybook-static
