+++ components/pages/submission-system/SchemaInvalidSubmissionNotification.tsx
@@ -5,15 +5,12 @@ import React from 'react';
 import { css } from '@emotion/core';
 import Link from 'uikit/Link';
 import Notification, { NOTIFICATION_INTERACTION_EVENTS } from 'uikit/notifications/Notification';
-import {
-  PROGRAM_CLINICAL_SUBMISSION_PATH,
-  PROGRAM_SHORT_NAME_PATH,
-  DOCS_DICTIONARY_PATH,
-} from 'global/constants/pages';
+import { PROGRAM_CLINICAL_SUBMISSION_PATH, PROGRAM_SHORT_NAME_PATH } from 'global/constants/pages';
 import SIDE_MENU_CLINICAL_SUBMISSION_STATE from './SIDE_MENU_CLINICAL_SUBMISSION_STATE.gql';
 import CLINICAL_SCHEMA_VERSION from './CLINICAL_SCHEMA_VERSION.gql';
 import { getConfig } from 'global/config';
 import urljoin from 'url-join';
+import { DOCS_DICTIONARY_PATH } from 'global/constants/docSitePaths';
 
 export const SchemaInvalidSubmissionNotification = ({
   marginTop,
