+++ global/constants/docSitePaths.tsx
@@ -1,0 +1,11 @@
+// docs paths
+export const DOCS_MANAGING_PROGRAM_ACCESS_PATH = `/docs/managing-program-access/`;
+export const DOCS_REGISTERING_SAMPLES_PATH = `/docs/registering-samples/`;
+export const DOCS_SUBMITTING_CLINICAL_DATA_PATH = `/docs/submitting-clinical-data/`;
+export const DOCS_SUBMITTED_DATA_PATH = `/docs/submitted-data/`;
+export const DOCS_SUBMITTING_MOLECULAR_DATA_PATH = `/docs/submitting-molecular-data/`;
+export const DOCS_DICTIONARY_PATH = `/dictionary/`;
+export const DOCS_API_TOKEN_PATH = `/docs/user-profile-and-api-token/`;
+export const DOCS_DATA_ACCESS_PATH = `/docs/data-access/`;
+export const DOCS_SUBMISSION_OVERVIEW_PATH = `/docs/submission-overview/`;
+export const DOCS_DATA_DOWNLOAD = `/docs/data-download/`;
